rbyd-rr: Trying another approach, 2-trunk diverging
This is a good checkpoint and is mostly working, though we're back to
recoloring the diverging alt black again. So no balance improvements.
But this already feels much better complexity-wise.

The fact that things could get back to a working state so quickly is a
good sign, or maybe just a sign I've been steeped in this algorithm for
too long...

---

The idea here is instead of a relatively complex 4-step state machine:

            diverged?                       diverged
  skip common -+-> write lower -> write common -> write upper -> done
               '-> write common -> done

We just write two trunks: one for the lower bound, one for the upper
bound.

We _do_ need to keep track of where we diverge so we can prune
correctly, so this is _technically_ still 4-steps, but it is at least
conceptually, and in code, much simpler:

             diverged?                       diverged
  write common -+-> write lower -> write common -> write upper -> done
                '-> done

Note that if we discover no tags in our range, we can terminate after
writing the lower/common trunk, which is nice. Previously we needed a
second pass.

The obvious downside is that we write the common trunk twice now. Which
is a bit of a downside, those alts will never really be used, but as a
tradeoff it really isn't that much of a waste. It's already possible for
range operations to need to write the full trunk twice, even for small
ranges:

         .-------o-------.
     .---o---.       .---o---.
   .-o-.   .-o-.   .-o-.   .-o-.
  .o. .o. .o. .o. .o. .o. .o. .o.
  a b c d e f g h i j k l m n o p
               '-+-'
               remove

The original motivation for trying yet-another-range-removal-algorithm
comes from attempting to solve issues with the color-balance of the
diverging alt.

The core conundrum being the case of two pending yellow splits. In our
previous algorithm, we only have one common trunk, so trying to
propagate two red edges violates tail recursion:

          .->                .->            .-> h=4 -.
    .-----b->              .-b->          .-b->      |
    |     .->              | .->          | .->      |
    | .---b->            .-y-b->        .-y-b->      |
    | |   .->            |   .->        |   .->      |
    | | .-b->            | .-b->        | .-b->      |
    | | | .->            b-b-b->      .-b-b-b->      |
  .-y-r-b-b-> rm me  =>           =>  |              +- unbalanced :(
  |       .->                         r-b-b-b->      |
  | .-----b->                           | | '->      |
  | |     .->                           | | .->      |
  | | .---b->                           | '-b->      |
  | | |   .->                           |   .->      |
  | | | .-b->                           | .-b->      |
  | | | | .->                           | | .->      |
  b-y-r-b-b->                           '-b-b-> h=3 -'
  ^                      ^
  diverging              lost color propagation

But if we have two trunks? Even only temporarily? This allows both
yellow splits/red edge propagation to settle tail recursively:

          .->                    .->              .-> h=3 -.
    .-----b->                  .-b->            .-b->      |
    |     .->                  | .->            | .->      |
    | .---b->                .-y-b->      .-----y-b->      |
    | |   .->                |   .->      |       .->      |
    | | .-b->                | .-b->      |     .-b->      |
    | | | .->            r---b-b-b->      | .---b-b->      |
  .-y-r-b-b-> rm me  =>  |            =>  | |              +- balanced :)
  |       .->            |       .->      y-r-b-b-b->      |
  | .-----b->            | .-----b->          | | '->      |
  | |     .->            | |     .->          | | .->      |
  | | .---b->            | | .---b->          | '-b->      |
  | | |   .->            | | |   .->          |   .->      |
  | | | .-b->            | | | .-b->          | .-b->      |
  | | | | .->            | | | | .->          | | .->      |
  b-y-r-b-b->            '-y-r-b-b->          '-b-b-> h=3 -'
  ^
  diverging

It's interesting to note while we _are_ currently recoloring the
diverging alt black (intentionally simplifying to algorithm to get
things moving), we are already allowing yellow splits/red edge
propagation by just writing out both trunks normally.

And that's what makes this yet-another-range-removal-algorithm appealing
and worth yet another iteration, the diverging trunks are no longer such
special cases. Not only will this make a better diverging color-balance
possible, it will hopefully make the whole diverging algorithm simpler,
easier, and cheaper. And it is already showing good signs so far.