rbyd-rr: Implemented very ugly, but working! diverging 2-3 nodes
It's a mess, but all tests are passing.

We're still recoloring the diverging alt, so hopefully I won't need to
eat my words, but at least on paper this should be able to preserve
colors for all 2-3 permutations of the diverging alt.

The key observation here is that diverging 2-3 nodes have three possible
permutations:

1. Diverging on the black alt:

         .->                .->          .->
     .---b->            .---b->      .---b->
     |   .->        =>  r-b-b->  =>  | .-b->
     | .-b-> rm me        |          | |
     | | .-> rm me        | .->      | |
     r-b-b->              '-b->      r-b-b->
       ^                  ^
       diverging          diverging

2. Diverging on the red alt:

         .->            r-b-b->        .--->
     .---b-> rm me      | |            |
     |   .-> rm me  =>  | | .->  =>    |
     | .-b->            | '-b->      r-b-b->
     | | .->            |   .->      |   .->
     r-b-b->            '---b->      '---b->
     ^                    ^
     diverging            diverging

3. Diverging on both alts:

         .->            b---b->      .---b->
     .---b-> rm me      |            |
     |   .-> rm me  =>  |        =>  |
     | .-b-> rm me      |            |
     | | .-> rm me      |   .->      |
     r-b-b->            '---b->      b---b->
     ^^^                ^
     diverging          diverging

With 3., both diverging, being the tricky one, where we need to both
switch to the diverged state while also collapsing the 3-node into a
2-node.

1. and 2. can both be deduplicated with a well-timed flip, but so far
it seems like 3. needs its own special case. At least these can all be
contained as extra conditions in the diverging alt logic, reducing the
possible states.

lfs_rbyd_appendattr is a complete mess now, and a lot of the diverging
logic is duplicated everywhere, but at least things seem to be working.