rbyd-rr: Rearranged diverged pruning/trimming after flipping
This was a bit more tricky than the other eager-flip related
transformations, mainly because we have to be careful to not prune the
diverging alt that connects the two diverged trunks. The diverging alt,
i.e. the first alt that diverges, passes all the criteria for pruning,
but is a bit special in that we need to keep it around until we stitch
the trunks together.

I ended up more-or-less just reverting the handling of both-diverging
nodes to being collapsed as a special case of our first encounter with
the diverging alt. Because we eagerly prune, both-diverging nodes can
only happen if they include the diverging alt. We can leveraging this to
simplify our diverging logic a bit, which is already crazy complicated.

Not only does this finish moving all of the alt-related logic into
"flipped space", it also moves all of the diverging logic together,
which is more readable and hopefully leads to better code deduplication
by the compiler.

Long story short, more code savings!

           code          stack
  before: 34244           2864
  after:  34176 (-0.2%)   2864 (+0.0%)

                     code          frame          stack
  appendattr before: 2232            216            568
  appendattr after:  2162 (-3.1%)    208 (-3.7%)    560 (-1.4%)

---

All of these code savings are making our 2-trunk range removal algorithm
more appealing:

                   code          stack
  rr-div-naive:   33968           2864
  rr-div-altn:    34304 (+1.0%)   2864 (+0.0%)
  rr-2trunk-altn: 34176 (+0.6%)   2864 (+0.0%)

                             code           frame           stack
  appendattr rr-stitching:   1940             184             536
  appendattr rr-div-naive:   2028 (+4.5%)     200 (+8.7%)     552 (+3.0%)
  appendattr rr-div-altn:    2198 (+13.3%)    216 (+17.4%)    568 (+6.0%)
  appendattr rr-2trunk-altn: 2162 (+11.4%)    208 (+13.0%)    560 (+4.5%)

That being said, it is getting increasingly hard to compare these
functions. You could argue the eager-flip transformations would also
result in code savings for the earlier iterations of our algorithm,
but it is worth noting the 2-trunk approach _did_ require more flips to
get working, so...