rbyd-rr: Explicitly deduplicated diverging conditions
I'm not really sure why the compiler isn't taking care of this for us.

Usually I prefer duplicated logic over more variables since it means
less state to keep track of when reading/debugging, and the compiler
will optimize it away anyways. But I guess these conditions are just too
complicated in this case?

Maybe the compiler is trying to take advantage of &&/|| short-circuiting
even with -Os?

Even marking the lfsr_tag_diverging* functions with
__attribute__((noinline, pure, const)) doesn't help...

Oh well, this is a case where we can just make the deduplication
explicit for a bit of code savings:

           code          stack
  before: 34176           2864
  after:  34080 (-0.3%)   2864 (+0.0%)

                     code          frame          stack
  appendattr before: 2162            208            560
  appendattr after:  2104 (-2.7%)    216 (+3.8%)    568 (+1.4%)