Trying to handle ecksums correctly when erased=>LFS_ERR_CORRUPT
It should be legal for block devices to return LFS_ERR_CORRUPT when
erased, this is common on devices with ECC, where the erased-state is
not valid ECC and results in LFS_ERR_CORRUPT.

If anything this is a better indicator than fixed-value erased-state,
but we need to make sure we track this with our ecksums consistently.
This gets a bit arbitrary.

Normally:

  valid = m[0] & 0x80
  cksum = crc32c(m)

If bd_read returns LFS_ERR_CORRUPT:

  valid = 0 & 0x80
  cksum = crc32c([])

Yeah, implementing this gets a bit funky, but the code cost is trivial:

           code          stack
  before: 33924           2824
  after:  33928 (+0.0%)   2824 (+0.0%)

Note this is only best effort right now, we really need tests over
erased=>LFS_ERR_CORRUPT...